pyproject.toml
[project]
name = "bob"
version = "0.1.0"
description = "Bob: AI-Powered Learning and Productivity Portal"
authors = [
  { name = "Iwan van der Kleijn", email = "iwanvanderkleijn@gmail.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.95.0,<1.0.0",
  "uvicorn[standard]>=0.22.0,<1.0.0",
  "jinja2>=3.1.2,<4.0.0",
  "pydantic>=1.10.11,<2.0.0",
  "chromadb>=0.3.31,<1.0.0",
  "langchain>=0.0.300,<1.0.0",
  "openai>=0.28.0,<1.0.0",
  "python-docx>=0.8.11,<1.0.0",
  "pdfplumber>=0.10.0,<1.0.0",
  "python-multipart>=0.0.20",
  "sqlalchemy>=1.4,<2.0",
  "python-dotenv>=1.0.0",
  "itsdangerous>=2.2.0",
  "bleach>=6.0.0,<7.0.0",
  "markupsafe>=2.1.0,<3.0.0",
  "redis>=6.2.0",
  "aiosqlite>=0.21.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0,<8.0.0",
  "black>=24.3.0,<25.0.0",
  "isort>=5.12.0,<6.0.0",
  "mypy>=1.5.1,<2.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bob"]
